+++ cmd/influx/pkg.go
@@ -369,6 +369,7 @@ func (b *cmdPkgBuilder) cmdStack() *cobra.Command {
 	cmd.AddCommand(
 		b.cmdStackInit(),
 		b.cmdStackList(),
+		b.cmdStackRemove(),
 	)
 	return cmd
 }
@@ -493,6 +494,99 @@ func (b *cmdPkgBuilder) stackListRunEFn(cmd *cobra.Command, args []string) error
 			"URLs":          stack.URLs,
 			"Created At":    stack.CreatedAt,
 		})
+	}
+
+	return nil
+}
+
+func (b *cmdPkgBuilder) cmdStackRemove() *cobra.Command {
+	cmd := b.newCmd("remove [--stack-id=ID1 --stack-id=ID2]", b.stackRemoveRunEFn, true)
+	cmd.Short = "Remove a stack(s) and all associated resources"
+	cmd.Aliases = []string{"rm", "uninstall"}
+
+	cmd.Flags().StringArrayVar(&b.stackIDs, "stack-id", nil, "Stack IDs to be removed")
+	cmd.MarkFlagRequired("stack-id")
+	registerPrintOptions(cmd, &b.hideHeaders, &b.json)
+
+	b.org.register(cmd, false)
+
+	return cmd
+}
+
+func (b *cmdPkgBuilder) stackRemoveRunEFn(cmd *cobra.Command, args []string) error {
+	pkgSVC, orgSVC, err := b.svcFn()
+	if err != nil {
+		return err
+	}
+
+	orgID, err := b.org.getID(orgSVC)
+	if err != nil {
+		return err
+	}
+
+	var stackIDs []influxdb.ID
+	for _, rawID := range b.stackIDs {
+		id, err := influxdb.IDFromString(rawID)
+		if err != nil {
+			return err
+		}
+		stackIDs = append(stackIDs, *id)
+	}
+
+	stacks, err := pkgSVC.ListStacks(context.Background(), orgID, pkger.ListFilter{
+		StackIDs: stackIDs,
+	})
+	if err != nil {
+		return err
+	}
+
+	printStack := func(stack pkger.Stack) error {
+		if b.json {
+			return b.writeJSON(stack)
+		}
+
+		tabW := b.newTabWriter()
+		defer func() {
+			tabW.Flush()
+			// add a breather line between confirm and printout
+			fmt.Fprintln(b.w)
+		}()
+
+		tabW.HideHeaders(b.hideHeaders)
+
+		tabW.WriteHeaders("ID", "OrgID", "Name", "Description", "Num Resources", "URLs", "Created At")
+		tabW.Write(map[string]interface{}{
+			"ID":            stack.ID,
+			"OrgID":         stack.OrgID,
+			"Name":          stack.Name,
+			"Description":   stack.Description,
+			"Num Resources": len(stack.Resources),
+			"URLs":          stack.URLs,
+			"Created At":    stack.CreatedAt,
+		})
+
+		return nil
+	}
+
+	for _, stack := range stacks {
+		if err := printStack(stack); err != nil {
+			return err
+		}
+
+		msg := fmt.Sprintf("Confirm removal of the stack[%s] and all associated resources (y/n)", stack.ID)
+		confirm := b.getInput(msg, "n")
+		if confirm != "y" {
+			continue
+		}
+
+		err := pkgSVC.DeleteStack(context.Background(), struct{ OrgID, UserID, StackID influxdb.ID }{
+			OrgID:   orgID,
+			UserID:  0,
+			StackID: stack.ID,
+		})
+		if err != nil {
+			return err
+		}
 	}
 
 	return nil